+++ CHANGELOG.md
@@ -5,6 +5,7 @@ principles.
 ## unreleased
 
 - Changed: passed the media api url from outside the js and vue scripts so the urls are in line with the rest of the admin
+- Fixed: removing a quill editor text field in pagebuilder is now possible
 
 ## 0.4.11 - 2020-03-18
 - Added: generate sitemap command and schedule